Move all birth date and SSN format stuff from locales.ts into its own file - birthdateAndSsnFormats.ts (in same folder)
