+++ src/client/l10n/locales.ts
@@ -1,4 +1,9 @@
 import { Locale as IsoLocale } from 'data/graphql'
+import {
+  ssnLengths,
+  ssnFormats,
+  birthdateFormats,
+} from './birthdateAndSsnFormats'
 
 export type LocaleLabel = 'se' | 'se-en' | 'no' | 'no-en' | 'dk' | 'dk-en'
 
@@ -8,45 +13,19 @@ export type LocaleData = {
   marketLabel: MarketLabel
   htmlLang: 'en' | 'sv' | 'no' | 'da'
   adtractionScript?: string
-  ssn: Ssn
-  birthdate: Birthdate
+  ssn: {
+    length: number
+    formatExample: string
+    formatRegex: RegExp
+  }
+  birthdate: {
+    formatExample: string
+    formatRegex: RegExp
+    backendFormatExample: string
+  }
 }
 
-type MarketLabel = 'SE' | 'NO' | 'DK'
-
-type Ssn = {
-  length: number
-  formatExample: string
-  formatRegex: RegExp
-}
-
-type Birthdate = {
-  formatExample: string
-  formatRegex: RegExp
-  backendFormatExample: string
-}
-
-const ssnFormats: Record<MarketLabel, RegExp> = {
-  SE: /^((19|20))([0-9]{2})(0[1-9]|1[0-2])(0[1-9]|[1-2][0-9]|3[0-1])([0-9]{4})$/,
-  NO: /^(0[1-9]|[1-2][0-9]|3[0-1])(0[1-9]|1[0-2])([0-9]{2})([0-9]{5})$/,
-  DK: /^(0[1-9]|[1-2][0-9]|3[0-1])(0[1-9]|1[0-2])([0-9]{2})([0-9]{4})$/,
-}
-
-const ssnLengths: Record<MarketLabel, number> = {
-  SE: 12,
-  NO: 11,
-  DK: 10,
-}
-
-const yearMonthDayFormat = /^(19[0-9]{2}|20[0-9]{2})-(0[1-9]|1[0-2])-(0[1-9]|[1-2][0-9]|3[0-1])$/
-const dayMonthYearFormat = /^(0[1-9]|[1-2][0-9]|3[0-1])-(0[1-9]|1[0-2])-(19[0-9]{2}|20[0-9]{2})$/
-
-export const birthdateFormats: Record<MarketLabel | 'default', RegExp> = {
-  default: yearMonthDayFormat,
-  SE: yearMonthDayFormat,
-  NO: dayMonthYearFormat,
-  DK: dayMonthYearFormat,
-}
+export type MarketLabel = 'SE' | 'NO' | 'DK'
 
 export const locales: Record<LocaleLabel, LocaleData> = {
   se: {
